Implementation of option validation and default settings via Settings API
* Enhancements:
- If WP_DEBUG is on, a block with the current options will be shown on the settings pages
- Use of selected() and checked() where applicable
- Added error message for when meta description removal fails
- Updated Webmaster tools urls
- Bow out early from displaying the post/taxonomy metabox if the post/taxonomy is not public
- Small function call efficiency improvements
- Added option to add meta keywords to post type archives
- Added error message when user tries to restore to defaults a non-existent blog (multi-site)
- Trim extravars for clean permalink, so the input can be readable (with spaces around comma)
- Moved any output adjustments to the input validation

* Bug fixes:
- Fix bug in delete_sitemaps() - wrong retrieval of needed options
- Failed meta description removal would still change the relevant option as if it had succeeded
- Corrected adding of hidden fb_admins fields to social form
- Not all options where removed on uninstall (still incomplete, see @todo's )

* Consistency:
- Where I came across it anyway, changed inappropriate double quotes to single quotes
- Output escaping of the form generating functions
- Fixed a few spelling errors
- Renamed wpseo/tracking_popup to tracking_popup_done so it's clear what the option means
- Flattened wpseo/theme_check option array for easier validation

= To do =
Deal with version number
Double-check that I've caught each and every option in use
Double-check that the forms send in all options, if one is missed, fall back to $old value
Rewrite the wpseo_defaults() function
Rewrite & move the reset_defaults() function
Double-check the maybe_upgrade() function
Double-check option-import from other plugins

+++ uninstall.php
@@ -3,11 +3,15 @@
  * @package Internals
  *
  * Code used when the plugin is removed (not just deactivated but actively deleted through the WordPress Admin).
+ *
+ * @todo remove meta data for posts/pages ?
+ * @todo remove wpseo_taxonomy_meta option ?
+ * @todo deal with multisite uninstall ?
  */
 
-if( !defined( 'ABSPATH') && !defined('WP_UNINSTALL_PLUGIN') )
-    exit();
+if ( !current_user_can( 'activate_plugins' ) || ( !defined( 'ABSPATH' ) || !defined( 'WP_UNINSTALL_PLUGIN' ) ) )
+	exit();
 
-foreach ( array('wpseo', 'wpseo_indexation', 'wpseo_permalinks', 'wpseo_titles', 'wpseo_rss', 'wpseo_internallinks', 'wpseo_xml', 'wpseo_social') as $option) {
+foreach ( array( 'wpseo', 'wpseo_indexation', 'wpseo_permalinks', 'wpseo_titles', 'wpseo_rss', 'wpseo_internallinks', 'wpseo_xml', 'wpseo_social', 'wpseo_ms' ) as $option ) {
 	delete_option( $option );
 }
